Updated server src, unzipped .rars
updated server src:
map generator runs on server start. it takes about 3 minutes to generate a map with the default settings (with my processor) so don't be worried if it takes a few minutes. if it takes too long, download the source and in the Server class, look for the InitializeServerData method, and within, change WorldGenerator.RenderBitmap()'s parameters to something lower.
In the future I'm adding multithreading to the rendering process to make it more time efficient.

unzipped .rars to save myself some hassle when updating